--- a/rueidisprob/bloomfilter.go
+++ b/rueidisprob/bloomfilter.go
@@ -16,17 +16,38 @@ const (
 
 const (
 	addMultiScript = `
-local count = tonumber(ARGV[1])
+local hashIterations = tonumber(ARGV[1])
 local numElements = tonumber(#ARGV) - 1
 local filterKey = KEYS[1]
 local counterKey = KEYS[2]
 
-for i = 2, numElements + 1 do
-    local index = tonumber(ARGV[i])
-    redis.call('SETBIT', filterKey, index, 1)
+local bitfieldArgs = { filterKey }
+for i=2, numElements+1 do
+    table.insert(bitfieldArgs, 'SET')
+    table.insert(bitfieldArgs, 'u1')
+    table.insert(bitfieldArgs, ARGV[i])
+    table.insert(bitfieldArgs, '1')
 end
 
-return redis.call('INCRBY', counterKey, count)
+local setBits = redis.call('BITFIELD', unpack(bitfieldArgs))
+
+local counter = 0
+local oneBits = 0
+for i=1, #setBits do
+	if setBits[i] == 1 then
+		oneBits = oneBits + 1
+	end
+
+	if i % hashIterations == 0 then
+		if oneBits ~= hashIterations then
+			counter = counter + 1
+		end
+
+		oneBits = 0
+	end
+end
+
+return redis.call('INCRBY', counterKey, counter)
 `
 
 	existsMultiScript = `
@@ -34,12 +55,16 @@ local numElements = tonumber(#ARGV)
 local filterKey = KEYS[1]
 local result = {}
 
+local bitfieldArgs = { filterKey }
 for i=1, numElements do
 	local index = tonumber(ARGV[i])
-	result[i] = redis.call('GETBIT', filterKey, index)
+
+	table.insert(bitfieldArgs, 'GET')
+	table.insert(bitfieldArgs, 'u1')
+	table.insert(bitfieldArgs, index)
 end
 
-return result
+return redis.call('BITFIELD', unpack(bitfieldArgs))
 `
 
 	resetScript = `
@@ -109,10 +134,17 @@ type bloomFilter struct {
 	counter string
 
 	// hashIterations is the number of hash functions to use.
-	hashIterations uint
+	hashIterations      uint
+	hashIterationString string
 
 	// size is the number of bits to use.
 	size uint
+
+	addMultiScript *rueidis.Lua
+	addMultiKeys   []string
+
+	existsMultiScript *rueidis.Lua
+	existsMultiKeys   []string
 }
 
 // NewBloomFilter creates a new Bloom filter.
@@ -145,13 +177,19 @@ func NewBloomFilter(
 	hashIterations := numberOfHashFunctions(size, expectedNumberOfItems)
 
 	// NOTE: https://redis.io/docs/reference/cluster-spec/#hash-tags
-	namespace := "{" + name + "}"
+	bfName := "{" + name + "}"
+	counterName := bfName + ":c"
 	return &bloomFilter{
-		client:         client,
-		name:           namespace,
-		counter:        namespace + ":c",
-		hashIterations: hashIterations,
-		size:           size,
+		client:              client,
+		name:                bfName,
+		counter:             counterName,
+		hashIterations:      hashIterations,
+		hashIterationString: strconv.FormatUint(uint64(hashIterations), 10),
+		size:                size,
+		addMultiScript:      rueidis.NewLuaScript(addMultiScript),
+		addMultiKeys:        []string{bfName, counterName},
+		existsMultiScript:   rueidis.NewLuaScript(existsMultiScript),
+		existsMultiKeys:     []string{bfName},
 	}, nil
 }
 
@@ -178,19 +216,10 @@ func (c *bloomFilter) AddMulti(ctx context.Context, keys []string) error {
 	indexes := c.indexes(keys)
 
 	var args []string
-	args = append(args, strconv.Itoa(len(keys)))
+	args = append(args, c.hashIterationString)
 	args = append(args, indexes...)
 
-	resp := c.client.Do(
-		ctx,
-		c.client.B().
-			Eval().
-			Script(addMultiScript).
-			Numkeys(2).
-			Key(c.name, c.counter).
-			Arg(args...).
-			Build(),
-	)
+	resp := c.addMultiScript.Exec(ctx, c.client, c.addMultiKeys, args)
 	if resp.Error() != nil {
 		return resp.Error()
 	}
@@ -225,16 +254,7 @@ func (c *bloomFilter) ExistsMulti(ctx context.Context, keys []string) ([]bool, e
 
 	indexes := c.indexes(keys)
 
-	resp := c.client.Do(
-		ctx,
-		c.client.B().
-			Eval().
-			Script(existsMultiScript).
-			Numkeys(1).
-			Key(c.name).
-			Arg(indexes...).
-			Build(),
-	)
+	resp := c.existsMultiScript.Exec(ctx, c.client, c.existsMultiKeys, indexes)
 	if resp.Error() != nil {
 		return nil, resp.Error()
 	}
@@ -244,9 +264,23 @@ func (c *bloomFilter) ExistsMulti(ctx context.Context, keys []string) ([]bool, e
 		return nil, err
 	}
 
-	result := make([]bool, len(is))
+	result := make([]bool, len(keys))
+	resultIdx := 0
+	oneBits := 0
+	iterations := int(c.hashIterations)
 	for i, v := range is {
-		result[i] = v == 1
+		if v == 1 {
+			oneBits++
+		}
+
+		if (i+1)%iterations == 0 {
+			if oneBits == iterations {
+				result[resultIdx] = true
+			}
+
+			resultIdx++
+			oneBits = 0
+		}
 	}
 	return result, nil
 }