--- a/rueidisprob/go.mod
+++ b/rueidisprob/go.mod
@@ -1,11 +1,11 @@
 module github.com/redis/rueidis/rueidisprob
 
-go 1.22.0
+go 1.20.0
 
 replace github.com/redis/rueidis => ../
 
 require (
-	github.com/redis/rueidis v0.0.0-00010101000000-000000000000
+	github.com/redis/rueidis v1.0.32
 	github.com/twmb/murmur3 v1.1.8
 )