--- a/rueidisprob/index.go
+++ b/rueidisprob/index.go
@@ -15,6 +15,10 @@ func indexes(data []byte, iteration, maxSize uint) []uint64 {
 	return indices
 }
 
+func hash(data []byte) (uint64, uint64) {
+	return murmur3.Sum128(data)
+}
+
 func index(h1, h2 uint64, i uint, maxSize uint64) uint64 {
 	offset := h1 + uint64(i)*h2
 	return offset % maxSize