discussion: https://github.com/redis/rueidis/discussions/486

standard bloom filter for `rueidisprob`