@rueian 2024-03-10
Hi @proost,

There are a couple of things we can address and discuss about:

1. Please add a test entry to the .circleci/config.yml, for example https://github.com/redis/rueidis/blob/b08dc62ed45c49f607e44d3af5f9bfc4e92fe21f/.circleci/config.yml#L61-L67
2. The filter key and the counter key should be located in the same Redis slot by using a [tag](https://redis.io/docs/reference/cluster-spec/#hash-tags) so that these Lua scripts can work in cluster mode.
3. The current implementation increments the counter no matter what, but that is inaccurate. Can we check the SETBIT results before incrementing the count? In other words, if the sum of the return values of SETBIT equals to the number of hash iterations, we should not increment the counter.
4. We use SETBIT one by one for each index. Can we make it a batch for a key with BITFIELD + multiple SETs?
5. We should not use EVAL every time. That hurts performance. Can we leverage the `rueidis.NewLuaScript`?