@rueian 2024-03-14
Hi @proost, thanks! This looks good to me. I also made a small benchmark between the latest commit and the previous commit d118842c3649:

```
goos: darwin
goarch: arm64
pkg: cmd/bloom
    │ d118842c3649    │        feat-rueidisprob         │
    │     sec/op      │   sec/op     vs base            │
-10       48.08µ ± 0%   48.12µ ± 0%  ~ (p=0.183 n=20)

    │ d118842c3649    │            feat-rueidisprob            │
    │      B/op       │     B/op      vs base                  │
-10      8.749Ki ± 1%   3.621Ki ± 0%  -58.61% (p=0.000 n=20)

    │ d118842c3649     │           feat-rueidisprob           │
    │    allocs/op     │ allocs/op   vs base                  │
-10        90.00 ± 0%   74.00 ± 0%  -17.78% (p=0.000 n=20)
```

The CPU loading of the Redis server in these two cases was both ~97% and ~98%. The latest feat-rueidisprob had slightly heavy CPU loading on the Redis, but it had -58.61% and -17.78% reduction in `B/op` and `allocs/op`.

Benchmark Code:
```go
package main

import (
	"context"
	"math/rand/v2"
	"strconv"
	"testing"

	"github.com/redis/rueidis"
	"github.com/redis/rueidis/rueidisprob"
)

func Benchmark(b *testing.B) {
	c, err := rueidis.NewClient(rueidis.ClientOption{InitAddress: []string{"127.0.0.1:6379"}})
	if err != nil {
		panic(err)
	}
	defer c.Close()

	bf, err := rueidisprob.NewBloomFilter(c, "bf", 1000000, 0.01)
	if err != nil {
		panic(err)
	}

	if err := bf.Add(context.Background(), "1"); err != nil {
		panic(err)
	}

	keys := make([]string, 10)
	for i := range keys {
		keys[i] = strconv.Itoa(rand.IntN(10000))
	}

	b.ReportAllocs()
	b.ResetTimer()
	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			if _, err := bf.ExistsMulti(context.Background(), keys); err != nil {
				panic(err)
			}
		}
	})
	b.StopTimer()
}

```