@proost 2024-03-15
@rueian 
[a2df9bb](https://github.com/redis/rueidis/pull/493/commits/a2df9bba11cefe16b578f7e9ae2847cc660dfc35)
I want to add this package to main README file. is it ok?