@proost 2024-03-17
@rueian 
Thank you for accepting proposal! I will send Counting Bloom Filter Design to Discussion soon.